Refactor constraints to provide expressions
This decouples constraints from residual calculation, allowing us to
introduce constraints that provide more than one residual. Parts of the
code working at the residual-level, like the optimizers, can now
continue working on a flat list of expressions, rather than operating on
a list of constraints with differing valency.
