+++ fiksi/src/constraints/mod.rs
@@ -3,29 +3,11 @@
 
 //! Constraints between geometric elements.
 
-#[cfg(not(feature = "std"))]
-use crate::floatfuncs::FloatFuncs;
-
 use crate::{
-    ConstraintHandle, ElementHandle, EncodedConstraint, EncodedElement, Subsystem, System,
-    elements, graph::IncidentElements,
+    ConstraintHandle, ElementHandle, EncodedElement, System, elements, graph::IncidentElements,
 };
 
-trait FloatExt {
-    /// Returns the square of `self`.
-    ///
-    /// Using `std`, you'd be able to do this using `self.powi(2)`, and have this be compiled to a
-    /// `self * self`. However, when compiling using `libm`, there is no `powi` and libm's
-    /// `self.powf(2.)` doesn't compile away.
-    fn square(self) -> Self;
-}
-
-impl FloatExt for f64 {
-    #[inline(always)]
-    fn square(self) -> Self {
-        self * self
-    }
-}
+pub(crate) mod expressions;
 
 pub(crate) mod constraint {
     use core::marker::PhantomData;
@@ -85,8 +67,10 @@ pub(crate) mod constraint {
                 "Tried to evaluate a constraint that is not part of this `System`"
             );
 
+            // TODO: handle constraints with `valency > 1`
             let constraint = &system.constraints[self.id as usize];
-            utils::calculate_residual(constraint, &system.variables)
+            let expression = &system.expressions[constraint.expressions_idx as usize];
+            utils::calculate_residual(expression, &system.variables)
         }
 
         /// Get a type-erased handle to the constraint.
@@ -125,8 +109,10 @@ pub(crate) mod constraint {
                 "Tried to get a constraint that is not part of this `System`"
             );
 
+            // TODO: handle constraints with `valency > 1`
             let constraint = &system.constraints[self.id as usize];
-            utils::calculate_residual(constraint, &system.variables)
+            let expression = &system.expressions[constraint.expressions_idx as usize];
+            utils::calculate_residual(expression, &system.variables)
         }
 
         /// Get a typed handle to the constraint.
@@ -234,13 +220,7 @@ pub(crate) mod constraint {
 }
 
 /// Constrain two points to have a given straight-line distance between each other.
-pub struct PointPointDistance {
-    point1_idx: u32,
-    point2_idx: u32,
-
-    /// Euclidean distance.
-    distance: f64,
-}
+pub struct PointPointDistance {}
 
 impl sealed::ConstraintInner for PointPointDistance {
     fn tag() -> ConstraintTag {
@@ -267,111 +247,24 @@ impl PointPointDistance {
             unreachable!()
         };
 
-        let constraint = Self {
-            point1_idx,
-            point2_idx,
-            distance,
-        };
-
         system.graph.add_constraint(
             1,
             IncidentElements::from_array([point1.drop_system_id(), point2.drop_system_id()]),
         );
-        system.add_constraint(EncodedConstraint::PointPointDistance(constraint))
-    }
-
-    /// If only, say, the residual or some subset of the gradient entries are actually used, and
-    /// that is clear from the code at the call site, the compiler should be able to correctly
-    /// remove the dead calculations as this is marked as `#[inline(always)]`. That allows us not to
-    /// to duplicate code unnecessarily for the calculations. Plus, calculating the residuals and
-    /// the Jacobian at the same time is a common case, and for some constraints it's more efficient
-    /// when they're calculated together.
-    #[inline(always)]
-    fn compute_residual_and_gradient_(
-        variables: &[f64; 4],
-        param_distance: f64,
-    ) -> (f64, [f64; 4]) {
-        let point1 = kurbo::Point {
-            x: variables[0],
-            y: variables[1],
-        };
-        let point2 = kurbo::Point {
-            x: variables[2],
-            y: variables[3],
-        };
-
-        let distance = ((point1.x - point2.x).square() + (point1.y - point2.y).square()).sqrt();
-        let residual = distance - param_distance;
-
-        let distance_recip = 1. / distance;
-        let gradient = [
-            (point1.x - point2.x) * distance_recip,
-            (point1.y - point2.y) * distance_recip,
-            -(point1.x - point2.x) * distance_recip,
-            -(point1.y - point2.y) * distance_recip,
-        ];
-
-        (residual, gradient)
-    }
-
-    pub(crate) fn compute_residual(&self, variables: &[f64]) -> f64 {
-        // The compiler should be able to optimize this such that only the residual is calculated.
-        // See the note about inlining on [`PointPointDistance::compute_residual_and_gradient_`].
-        Self::compute_residual_and_gradient_(
-            &[
-                variables[self.point1_idx as usize],
-                variables[self.point1_idx as usize + 1],
-                variables[self.point2_idx as usize],
-                variables[self.point2_idx as usize + 1],
-            ],
-            self.distance,
+        system.add_constraint(
+            ConstraintTag::PointPointDistance,
+            [expressions::PointPointDistance {
+                point1_idx,
+                point2_idx,
+                distance,
+            }
+            .into()],
         )
-        .0
-    }
-
-    pub(crate) fn compute_residual_and_gradient(
-        &self,
-        subsystem: &Subsystem<'_>,
-        variables: &[f64],
-        residual: &mut f64,
-        gradient: &mut [f64],
-    ) {
-        let (r, g) = Self::compute_residual_and_gradient_(
-            &[
-                variables[self.point1_idx as usize],
-                variables[self.point1_idx as usize + 1],
-                variables[self.point2_idx as usize],
-                variables[self.point2_idx as usize + 1],
-            ],
-            self.distance,
-        );
-
-        *residual += r;
-
-        if let Some(idx) = subsystem.free_variable_index(self.point1_idx) {
-            gradient[idx as usize] += g[0];
-        }
-        if let Some(idx) = subsystem.free_variable_index(self.point1_idx + 1) {
-            gradient[idx as usize] += g[1];
-        }
-        if let Some(idx) = subsystem.free_variable_index(self.point2_idx) {
-            gradient[idx as usize] += g[2];
-        }
-        if let Some(idx) = subsystem.free_variable_index(self.point2_idx + 1) {
-            gradient[idx as usize] += g[3];
-        }
     }
 }
 
 /// Constrain three points to describe a given angle.
-pub struct PointPointPointAngle {
-    point1_idx: u32,
-    point2_idx: u32,
-    point3_idx: u32,
-
-    /// Angle in radians.
-    angle: f64,
-}
+pub struct PointPointPointAngle {}
 
 impl sealed::ConstraintInner for PointPointPointAngle {
     fn tag() -> ConstraintTag {
@@ -403,13 +296,6 @@ impl PointPointPointAngle {
             unreachable!()
         };
 
-        let constraint = Self {
-            point1_idx,
-            point2_idx,
-            point3_idx,
-            angle,
-        };
-
         system.graph.add_constraint(
             1,
             IncidentElements::from_array([
@@ -418,118 +304,16 @@ impl PointPointPointAngle {
                 point3.drop_system_id(),
             ]),
         );
-        system.add_constraint(EncodedConstraint::PointPointPointAngle(constraint))
-    }
-
-    // See the note about inlining on [`PointPointDistance::compute_residual_and_gradient_`].
-    #[inline(always)]
-    fn compute_residual_and_gradient_(variables: &[f64; 6], param_angle: f64) -> (f64, [f64; 6]) {
-        let point1 = kurbo::Point {
-            x: variables[0],
-            y: variables[1],
-        };
-        let point2 = kurbo::Point {
-            x: variables[2],
-            y: variables[3],
-        };
-        let point3 = kurbo::Point {
-            x: variables[4],
-            y: variables[5],
-        };
-
-        let u = point1 - point2;
-        let v = point3 - point2;
-
-        let angle = v.atan2() - u.atan2();
-        let angle = if angle > core::f64::consts::PI {
-            angle - 2.0 * core::f64::consts::PI
-        } else if angle < -core::f64::consts::PI {
-            angle + 2.0 * core::f64::consts::PI
-        } else {
-            angle
-        };
-
-        let residual = angle - param_angle;
-
-        let u_squared_recip = u.length_squared().recip();
-        let v_squared_recip = v.length_squared().recip();
-
-        let dangle_dpoint1x = u.y * u_squared_recip;
-        let dangle_dpoint1y = -u.x * u_squared_recip;
-        let dangle_dpoint3x = -v.y * v_squared_recip;
-        let dangle_dpoint3y = v.x * v_squared_recip;
-
-        let dangle_dpoint2x = -dangle_dpoint1x - dangle_dpoint3x;
-        let dangle_dpoint2y = -dangle_dpoint1y - dangle_dpoint3y;
-
-        let gradient = [
-            dangle_dpoint1x,
-            dangle_dpoint1y,
-            dangle_dpoint2x,
-            dangle_dpoint2y,
-            dangle_dpoint3x,
-            dangle_dpoint3y,
-        ];
-
-        (residual, gradient)
-    }
-
-    pub(crate) fn compute_residual(&self, variables: &[f64]) -> f64 {
-        // The compiler should be able to optimize this such that only the residual is calculated.
-        // See the note about inlining on [`PointPointDistance::compute_residual_and_gradient_`].
-        Self::compute_residual_and_gradient_(
-            &[
-                variables[self.point1_idx as usize],
-                variables[self.point1_idx as usize + 1],
-                variables[self.point2_idx as usize],
-                variables[self.point2_idx as usize + 1],
-                variables[self.point3_idx as usize],
-                variables[self.point3_idx as usize + 1],
-            ],
-            self.angle,
+        system.add_constraint(
+            ConstraintTag::PointPointPointAngle,
+            [expressions::PointPointPointAngle {
+                point1_idx,
+                point2_idx,
+                point3_idx,
+                angle,
+            }
+            .into()],
         )
-        .0
-    }
-
-    pub(crate) fn compute_residual_and_gradient(
-        &self,
-        subsystem: &Subsystem<'_>,
-        variables: &[f64],
-        residual: &mut f64,
-        gradient: &mut [f64],
-    ) {
-        let (r, g) = Self::compute_residual_and_gradient_(
-            &[
-                variables[self.point1_idx as usize],
-                variables[self.point1_idx as usize + 1],
-                variables[self.point2_idx as usize],
-                variables[self.point2_idx as usize + 1],
-                variables[self.point3_idx as usize],
-                variables[self.point3_idx as usize + 1],
-            ],
-            self.angle,
-        );
-
-        *residual += r;
-
-        if let Some(idx) = subsystem.free_variable_index(self.point1_idx) {
-            gradient[idx as usize] += g[0];
-        }
-        if let Some(idx) = subsystem.free_variable_index(self.point1_idx + 1) {
-            gradient[idx as usize] += g[1];
-        }
-        if let Some(idx) = subsystem.free_variable_index(self.point2_idx) {
-            gradient[idx as usize] += g[2];
-        }
-        if let Some(idx) = subsystem.free_variable_index(self.point2_idx + 1) {
-            gradient[idx as usize] += g[3];
-        }
-        if let Some(idx) = subsystem.free_variable_index(self.point3_idx) {
-            gradient[idx as usize] += g[4];
-        }
-        if let Some(idx) = subsystem.free_variable_index(self.point3_idx + 1) {
-            gradient[idx as usize] += g[5];
-        }
     }
 }
 
@@ -538,11 +322,7 @@ impl PointPointPointAngle {
 /// Note this does not constrain the point to lie on the line *segment* defined by `line`. This is
 /// equivalent to constraining the three points (the two points of the line and the point proper)
 /// to be collinear.
-pub struct PointLineIncidence {
-    point_idx: u32,
-    line_point1_idx: u32,
-    line_point2_idx: u32,
-}
+pub struct PointLineIncidence {}
 
 impl sealed::ConstraintInner for PointLineIncidence {
     fn tag() -> ConstraintTag {
@@ -569,12 +349,6 @@ impl PointLineIncidence {
             unreachable!()
         };
 
-        let constraint = Self {
-            point_idx,
-            line_point1_idx,
-            line_point2_idx,
-        };
-
         system.graph.add_constraint(
             1,
             IncidentElements::from_array([
@@ -583,105 +357,20 @@ impl PointLineIncidence {
                 system.variable_to_primitive[line_point2_idx as usize],
             ]),
         );
-        system.add_constraint(EncodedConstraint::PointLineIncidence(constraint))
-    }
-
-    // See the note about inlining on [`PointPointDistance::compute_residual_and_gradient_`].
-    #[inline(always)]
-    fn compute_residual_and_gradient_(variables: &[f64; 6]) -> (f64, [f64; 6]) {
-        let point1 = kurbo::Point {
-            x: variables[0],
-            y: variables[1],
-        };
-        let point2 = kurbo::Point {
-            x: variables[2],
-            y: variables[3],
-        };
-        let point3 = kurbo::Point {
-            x: variables[4],
-            y: variables[5],
-        };
-
-        // For collinear points, the triangle defined by those points has area 0.
-        let residual = point1.x * (point2.y - point3.y)
-            + point2.x * (point3.y - point1.y)
-            + point3.x * (point1.y - point2.y);
-
-        let gradient = [
-            point2.y - point3.y,
-            -point2.x + point3.x,
-            point3.y - point1.y,
-            point1.x - point3.x,
-            point1.y - point2.y,
-            -point1.x + point2.x,
-        ];
-
-        (residual, gradient)
-    }
-
-    pub(crate) fn compute_residual(&self, variables: &[f64]) -> f64 {
-        // The compiler should be able to optimize this such that only the residual is calculated.
-        // See the note about inlining on [`PointPointDistance::compute_residual_and_gradient_`].
-        Self::compute_residual_and_gradient_(&[
-            variables[self.point_idx as usize],
-            variables[self.point_idx as usize + 1],
-            variables[self.line_point1_idx as usize],
-            variables[self.line_point1_idx as usize + 1],
-            variables[self.line_point2_idx as usize],
-            variables[self.line_point2_idx as usize + 1],
-        ])
-        .0
-    }
-
-    pub(crate) fn compute_residual_and_gradient(
-        &self,
-        subsystem: &Subsystem<'_>,
-        variables: &[f64],
-        residual: &mut f64,
-        gradient: &mut [f64],
-    ) {
-        let (r, g) = Self::compute_residual_and_gradient_(&[
-            variables[self.point_idx as usize],
-            variables[self.point_idx as usize + 1],
-            variables[self.line_point1_idx as usize],
-            variables[self.line_point1_idx as usize + 1],
-            variables[self.line_point2_idx as usize],
-            variables[self.line_point2_idx as usize + 1],
-        ]);
-
-        *residual += r;
-
-        if let Some(idx) = subsystem.free_variable_index(self.point_idx) {
-            gradient[idx as usize] += g[0];
-        }
-        if let Some(idx) = subsystem.free_variable_index(self.point_idx + 1) {
-            gradient[idx as usize] += g[1];
-        }
-        if let Some(idx) = subsystem.free_variable_index(self.line_point1_idx) {
-            gradient[idx as usize] += g[2];
-        }
-        if let Some(idx) = subsystem.free_variable_index(self.line_point1_idx + 1) {
-            gradient[idx as usize] += g[3];
-        }
-        if let Some(idx) = subsystem.free_variable_index(self.line_point2_idx) {
-            gradient[idx as usize] += g[4];
-        }
-        if let Some(idx) = subsystem.free_variable_index(self.line_point2_idx + 1) {
-            gradient[idx as usize] += g[5];
-        }
+        system.add_constraint(
+            ConstraintTag::PointLineIncidence,
+            [expressions::PointLineIncidence {
+                point_idx,
+                line_point1_idx,
+                line_point2_idx,
+            }
+            .into()],
+        )
     }
 }
 
 /// Constrain two lines to describe a given angle.
-pub struct LineLineAngle {
-    line1_point1_idx: u32,
-    line1_point2_idx: u32,
-    line2_point1_idx: u32,
-    line2_point2_idx: u32,
-
-    /// Angle in radians.
-    angle: f64,
-}
+pub struct LineLineAngle {}
 
 impl sealed::ConstraintInner for LineLineAngle {
     fn tag() -> ConstraintTag {
@@ -712,14 +401,6 @@ impl LineLineAngle {
             unreachable!()
         };
 
-        let constraint = Self {
-            line1_point1_idx,
-            line1_point2_idx,
-            line2_point1_idx,
-            line2_point2_idx,
-            angle,
-        };
-
         system.graph.add_constraint(
             1,
             IncidentElements::from_array([
@@ -729,141 +410,22 @@ impl LineLineAngle {
                 system.variable_to_primitive[line2_point2_idx as usize],
             ]),
         );
-        system.add_constraint(EncodedConstraint::LineLineAngle(constraint))
-    }
-
-    // See the note about inlining on [`PointPointDistance::compute_residual_and_gradient_`].
-    #[inline(always)]
-    fn compute_residual_and_gradient_(variables: &[f64; 8], param_angle: f64) -> (f64, [f64; 8]) {
-        let line1_point1 = kurbo::Point {
-            x: variables[0],
-            y: variables[1],
-        };
-        let line1_point2 = kurbo::Point {
-            x: variables[2],
-            y: variables[3],
-        };
-        let line2_point1 = kurbo::Point {
-            x: variables[4],
-            y: variables[5],
-        };
-        let line2_point2 = kurbo::Point {
-            x: variables[6],
-            y: variables[7],
-        };
-
-        let u = line1_point2 - line1_point1;
-        let v = line2_point2 - line2_point1;
-
-        let angle = v.atan2() - u.atan2();
-        let angle = if angle > core::f64::consts::PI {
-            angle - 2.0 * core::f64::consts::PI
-        } else if angle < -core::f64::consts::PI {
-            angle + 2.0 * core::f64::consts::PI
-        } else {
-            angle
-        };
-
-        let residual = angle - param_angle;
-
-        let u_squared_recip = u.length_squared().recip();
-        let v_squared_recip = v.length_squared().recip();
-
-        let dangle_dline1_point1x = -u.y * u_squared_recip;
-        let dangle_dline1_point1y = u.x * u_squared_recip;
-        let dangle_dline2_point1x = v.y * v_squared_recip;
-        let dangle_dline2_point1y = -v.x * v_squared_recip;
-
-        let gradient = [
-            dangle_dline1_point1x,
-            dangle_dline1_point1y,
-            -dangle_dline1_point1x,
-            -dangle_dline1_point1y,
-            dangle_dline2_point1x,
-            dangle_dline2_point1y,
-            -dangle_dline2_point1x,
-            -dangle_dline2_point1y,
-        ];
-
-        (residual, gradient)
-    }
-
-    pub(crate) fn compute_residual(&self, variables: &[f64]) -> f64 {
-        // The compiler should be able to optimize this such that only the residual is calculated.
-        // See the note about inlining on [`PointPointDistance::compute_residual_and_gradient_`].
-        Self::compute_residual_and_gradient_(
-            &[
-                variables[self.line1_point1_idx as usize],
-                variables[self.line1_point1_idx as usize + 1],
-                variables[self.line1_point2_idx as usize],
-                variables[self.line1_point2_idx as usize + 1],
-                variables[self.line2_point1_idx as usize],
-                variables[self.line2_point1_idx as usize + 1],
-                variables[self.line2_point2_idx as usize],
-                variables[self.line2_point2_idx as usize + 1],
-            ],
-            self.angle,
+        system.add_constraint(
+            ConstraintTag::LineLineAngle,
+            [expressions::LineLineAngle {
+                line1_point1_idx,
+                line1_point2_idx,
+                line2_point1_idx,
+                line2_point2_idx,
+                angle,
+            }
+            .into()],
         )
-        .0
-    }
-
-    pub(crate) fn compute_residual_and_gradient(
-        &self,
-        subsystem: &Subsystem<'_>,
-        variables: &[f64],
-        residual: &mut f64,
-        gradient: &mut [f64],
-    ) {
-        let (r, g) = Self::compute_residual_and_gradient_(
-            &[
-                variables[self.line1_point1_idx as usize],
-                variables[self.line1_point1_idx as usize + 1],
-                variables[self.line1_point2_idx as usize],
-                variables[self.line1_point2_idx as usize + 1],
-                variables[self.line2_point1_idx as usize],
-                variables[self.line2_point1_idx as usize + 1],
-                variables[self.line2_point2_idx as usize],
-                variables[self.line2_point2_idx as usize + 1],
-            ],
-            self.angle,
-        );
-
-        *residual += r;
-
-        if let Some(idx) = subsystem.free_variable_index(self.line1_point1_idx) {
-            gradient[idx as usize] += g[0];
-        }
-        if let Some(idx) = subsystem.free_variable_index(self.line1_point1_idx + 1) {
-            gradient[idx as usize] += g[1];
-        }
-        if let Some(idx) = subsystem.free_variable_index(self.line1_point2_idx) {
-            gradient[idx as usize] += g[2];
-        }
-        if let Some(idx) = subsystem.free_variable_index(self.line1_point2_idx + 1) {
-            gradient[idx as usize] += g[3];
-        }
-        if let Some(idx) = subsystem.free_variable_index(self.line2_point1_idx) {
-            gradient[idx as usize] += g[4];
-        }
-        if let Some(idx) = subsystem.free_variable_index(self.line2_point1_idx + 1) {
-            gradient[idx as usize] += g[5];
-        }
-        if let Some(idx) = subsystem.free_variable_index(self.line2_point2_idx) {
-            gradient[idx as usize] += g[6];
-        }
-        if let Some(idx) = subsystem.free_variable_index(self.line2_point2_idx + 1) {
-            gradient[idx as usize] += g[7];
-        }
     }
 }
 
 /// Constrain two lines to be parallel to each other.
-pub struct LineLineParallelism {
-    line1_point1_idx: u32,
-    line1_point2_idx: u32,
-    line2_point1_idx: u32,
-    line2_point2_idx: u32,
-}
+pub struct LineLineParallelism {}
 
 impl sealed::ConstraintInner for LineLineParallelism {
     fn tag() -> ConstraintTag {
@@ -902,123 +464,21 @@ impl LineLineParallelism {
                 system.variable_to_primitive[line2_point2_idx as usize],
             ]),
         );
-        system.add_constraint(EncodedConstraint::LineLineParallelism(Self {
-            line1_point1_idx,
-            line1_point2_idx,
-            line2_point1_idx,
-            line2_point2_idx,
-        }))
-    }
-
-    // See the note about inlining on [`PointPointDistance::compute_residual_and_gradient_`].
-    #[inline(always)]
-    fn compute_residual_and_gradient_(variables: &[f64; 8]) -> (f64, [f64; 8]) {
-        let line1_point1 = kurbo::Point {
-            x: variables[0],
-            y: variables[1],
-        };
-        let line1_point2 = kurbo::Point {
-            x: variables[2],
-            y: variables[3],
-        };
-        let line2_point1 = kurbo::Point {
-            x: variables[4],
-            y: variables[5],
-        };
-        let line2_point2 = kurbo::Point {
-            x: variables[6],
-            y: variables[7],
-        };
-
-        let u = line1_point2 - line1_point1;
-        let v = line2_point2 - line2_point1;
-
-        let residual = v.cross(u);
-
-        let gradient = [
-            v.y,  // l1p1x
-            -v.x, // l1p1y
-            -v.y, // l1p2x
-            v.x,  // l1p2y
-            -u.y, // l2p1x
-            u.x,  // l2p1y
-            u.y,  // l2p2x
-            -u.x, // l2p2y
-        ];
-
-        (residual, gradient)
-    }
-
-    pub(crate) fn compute_residual(&self, variables: &[f64]) -> f64 {
-        // The compiler should be able to optimize this such that only the residual is calculated.
-        // See the note about inlining on [`PointPointDistance::compute_residual_and_gradient_`].
-        Self::compute_residual_and_gradient_(&[
-            variables[self.line1_point1_idx as usize],
-            variables[self.line1_point1_idx as usize + 1],
-            variables[self.line1_point2_idx as usize],
-            variables[self.line1_point2_idx as usize + 1],
-            variables[self.line2_point1_idx as usize],
-            variables[self.line2_point1_idx as usize + 1],
-            variables[self.line2_point2_idx as usize],
-            variables[self.line2_point2_idx as usize + 1],
-        ])
-        .0
-    }
-
-    pub(crate) fn compute_residual_and_gradient(
-        &self,
-        subsystem: &Subsystem<'_>,
-        variables: &[f64],
-        residual: &mut f64,
-        gradient: &mut [f64],
-    ) {
-        let (r, g) = Self::compute_residual_and_gradient_(&[
-            variables[self.line1_point1_idx as usize],
-            variables[self.line1_point1_idx as usize + 1],
-            variables[self.line1_point2_idx as usize],
-            variables[self.line1_point2_idx as usize + 1],
-            variables[self.line2_point1_idx as usize],
-            variables[self.line2_point1_idx as usize + 1],
-            variables[self.line2_point2_idx as usize],
-            variables[self.line2_point2_idx as usize + 1],
-        ]);
-
-        *residual += r;
-
-        if let Some(idx) = subsystem.free_variable_index(self.line1_point1_idx) {
-            gradient[idx as usize] += g[0];
-        }
-        if let Some(idx) = subsystem.free_variable_index(self.line1_point1_idx + 1) {
-            gradient[idx as usize] += g[1];
-        }
-        if let Some(idx) = subsystem.free_variable_index(self.line1_point2_idx) {
-            gradient[idx as usize] += g[2];
-        }
-        if let Some(idx) = subsystem.free_variable_index(self.line1_point2_idx + 1) {
-            gradient[idx as usize] += g[3];
-        }
-        if let Some(idx) = subsystem.free_variable_index(self.line2_point1_idx) {
-            gradient[idx as usize] += g[4];
-        }
-        if let Some(idx) = subsystem.free_variable_index(self.line2_point1_idx + 1) {
-            gradient[idx as usize] += g[5];
-        }
-        if let Some(idx) = subsystem.free_variable_index(self.line2_point2_idx) {
-            gradient[idx as usize] += g[6];
-        }
-        if let Some(idx) = subsystem.free_variable_index(self.line2_point2_idx + 1) {
-            gradient[idx as usize] += g[7];
-        }
+        system.add_constraint(
+            ConstraintTag::LineLineParallelism,
+            [expressions::LineLineParallelism {
+                line1_point1_idx,
+                line1_point2_idx,
+                line2_point1_idx,
+                line2_point2_idx,
+            }
+            .into()],
+        )
     }
 }
 
 /// Constrain a line and a circle such that the line is tangent on the circle.
-pub struct LineCircleTangency {
-    line_point1_idx: u32,
-    line_point2_idx: u32,
-    circle_center_idx: u32,
-    circle_radius_idx: u32,
-}
+pub struct LineCircleTangency {}
 
 impl sealed::ConstraintInner for LineCircleTangency {
     fn tag() -> ConstraintTag {
@@ -1049,13 +509,6 @@ impl LineCircleTangency {
             unreachable!()
         };
 
-        let constraint = Self {
-            line_point1_idx,
-            line_point2_idx,
-            circle_center_idx,
-            circle_radius_idx,
-        };
-
         system.graph.add_constraint(
             1,
             IncidentElements::from_array([
@@ -1065,122 +518,16 @@ impl LineCircleTangency {
                 circle.drop_system_id(),
             ]),
         );
-        system.add_constraint(EncodedConstraint::LineCircleTangency(constraint))
-    }
-
-    // See the note about inlining on [`PointPointDistance::compute_residual_and_gradient_`].
-    #[inline(always)]
-    fn compute_residual_and_gradient_(variables: &[f64; 7]) -> (f64, [f64; 7]) {
-        let line_point1 = kurbo::Point {
-            x: variables[0],
-            y: variables[1],
-        };
-        let line_point2 = kurbo::Point {
-            x: variables[2],
-            y: variables[3],
-        };
-        let circle_center = kurbo::Point {
-            x: variables[4],
-            y: variables[5],
-        };
-        let circle_radius = variables[6];
-
-        let length2 = line_point1.distance_squared(line_point2);
-        let length = length2.sqrt();
-
-        // TODO: better handle degenerate lines of length 0.
-        if length == 0. {
-            return (0., [0.; 7]);
-        }
-
-        let length_recip = 1. / length;
-        let signed_area = line_point1.x * (line_point2.y - circle_center.y)
-            + line_point2.x * (circle_center.y - line_point1.y)
-            + circle_center.x * (line_point1.y - line_point2.y);
-
-        // We are interested in the _unsigned_ area here, as it does not matter on which side of
-        // the line the circle center lies. That does mean there is a cusp when the circle
-        // center is exactly on the line.
-        let residual = length_recip * signed_area.abs() - circle_radius;
-
-        let sign = signed_area.signum();
-        let length3_recip = 1. / (length2 * length);
-        let gradient = [
-            sign * length3_recip
-                * (length2 * (line_point2.y - circle_center.y)
-                    + signed_area * (line_point2.x - line_point1.x)),
-            sign * length3_recip
-                * (length2 * (-line_point2.x + circle_center.x)
-                    + signed_area * (line_point2.y - line_point1.y)),
-            sign * length3_recip
-                * (length2 * (circle_center.y - line_point1.y)
-                    - signed_area * (line_point2.x - line_point1.x)),
-            sign * length3_recip
-                * (length2 * (line_point1.x - circle_center.x)
-                    - signed_area * (line_point2.y - line_point1.y)),
-            sign * length_recip * (line_point1.y - line_point2.y),
-            sign * length_recip * (-line_point1.x + line_point2.x),
-            -1.,
-        ];
-
-        (residual, gradient)
-    }
-
-    pub(crate) fn compute_residual(&self, variables: &[f64]) -> f64 {
-        // The compiler should be able to optimize this such that only the residual is calculated.
-        // See the note about inlining on [`PointPointDistance::compute_residual_and_gradient_`].
-        Self::compute_residual_and_gradient_(&[
-            variables[self.line_point1_idx as usize],
-            variables[self.line_point1_idx as usize + 1],
-            variables[self.line_point2_idx as usize],
-            variables[self.line_point2_idx as usize + 1],
-            variables[self.circle_center_idx as usize],
-            variables[self.circle_center_idx as usize + 1],
-            variables[self.circle_radius_idx as usize],
-        ])
-        .0
-    }
-
-    pub(crate) fn compute_residual_and_gradient(
-        &self,
-        subsystem: &Subsystem<'_>,
-        variables: &[f64],
-        residual: &mut f64,
-        gradient: &mut [f64],
-    ) {
-        let (r, g) = Self::compute_residual_and_gradient_(&[
-            variables[self.line_point1_idx as usize],
-            variables[self.line_point1_idx as usize + 1],
-            variables[self.line_point2_idx as usize],
-            variables[self.line_point2_idx as usize + 1],
-            variables[self.circle_center_idx as usize],
-            variables[self.circle_center_idx as usize + 1],
-            variables[self.circle_radius_idx as usize],
-        ]);
-
-        *residual += r;
-
-        if let Some(idx) = subsystem.free_variable_index(self.line_point1_idx) {
-            gradient[idx as usize] += g[0];
-        }
-        if let Some(idx) = subsystem.free_variable_index(self.line_point1_idx + 1) {
-            gradient[idx as usize] += g[1];
-        }
-        if let Some(idx) = subsystem.free_variable_index(self.line_point2_idx) {
-            gradient[idx as usize] += g[2];
-        }
-        if let Some(idx) = subsystem.free_variable_index(self.line_point2_idx + 1) {
-            gradient[idx as usize] += g[3];
-        }
-        if let Some(idx) = subsystem.free_variable_index(self.circle_center_idx) {
-            gradient[idx as usize] += g[4];
-        }
-        if let Some(idx) = subsystem.free_variable_index(self.circle_center_idx + 1) {
-            gradient[idx as usize] += g[5];
-        }
-        if let Some(idx) = subsystem.free_variable_index(self.circle_radius_idx) {
-            gradient[idx as usize] += g[6];
-        }
+        system.add_constraint(
+            ConstraintTag::LineCircleTangency,
+            [expressions::LineCircleTangency {
+                line_point1_idx,
+                line_point2_idx,
+                circle_center_idx,
+                circle_radius_idx,
+            }
+            .into()],
+        )
     }
 }
 
@@ -1195,15 +542,15 @@ pub(crate) enum ConstraintTag {
     LineCircleTangency,
 }
 
-impl<'a> From<&'a EncodedConstraint> for ConstraintTag {
-    fn from(edge: &'a EncodedConstraint) -> Self {
-        match edge {
-            EncodedConstraint::PointPointDistance { .. } => Self::PointPointDistance,
-            EncodedConstraint::PointPointPointAngle { .. } => Self::PointPointPointAngle,
-            EncodedConstraint::PointLineIncidence { .. } => Self::PointLineIncidence,
-            EncodedConstraint::LineLineAngle { .. } => Self::LineLineAngle,
-            EncodedConstraint::LineLineParallelism { .. } => Self::LineLineParallelism,
-            EncodedConstraint::LineCircleTangency { .. } => Self::LineCircleTangency,
+impl ConstraintTag {
+    pub(crate) fn valency(self) -> u8 {
+        match self {
+            Self::PointPointDistance => PointPointDistance::VALENCY,
+            Self::PointPointPointAngle => PointPointPointAngle::VALENCY,
+            Self::PointLineIncidence => PointLineIncidence::VALENCY,
+            Self::LineLineAngle => LineLineAngle::VALENCY,
+            Self::LineLineParallelism => LineLineParallelism::VALENCY,
+            Self::LineCircleTangency => LineCircleTangency::VALENCY,
         }
     }
 }
@@ -1218,237 +565,36 @@ pub(crate) mod sealed {
 ///
 /// These can be added to a [`System`].
 #[expect(private_bounds, reason = "Sealed inner trait")]
-pub trait Constraint: sealed::ConstraintInner {}
-
-impl Constraint for PointPointDistance {}
-impl Constraint for PointPointPointAngle {}
-impl Constraint for PointLineIncidence {}
-impl Constraint for LineLineAngle {}
-impl Constraint for LineLineParallelism {}
-impl Constraint for LineCircleTangency {}
-
-#[cfg(test)]
-mod tests {
-    use core::array;
-
-    use crate::Rng;
-
-    use super::*;
-
-    /// Generate an array of random floats between 0. and 1. inclusive.
-    fn next_f64s<const N: usize>(rng: &mut Rng) -> [f64; N] {
-        array::from_fn(|_| rng.next_f64())
-    }
-
-    /// Calculate component-wise `a + b`.
-    fn add<const N: usize>(mut a: [f64; N], b: [f64; N]) -> [f64; N] {
-        for n in 0..N {
-            a[n] += b[n];
-        }
-        a
-    }
-
-    /// Calculate the dot product of a and b, `sum_n(a[n] * b[n])`.
-    fn dot<const N: usize>(a: [f64; N], b: [f64; N]) -> f64 {
-        let mut dot = 0.;
-        for n in 0..N {
-            dot += a[n] * b[n];
-        }
-        dot
-    }
-
-    /// For a given residual r(variables) and small step `delta` we assume
-    /// `r(variables + delta) ~= r(variables) + dot(gradient, delta)`
-    /// where `gradient` is the vector of partial derivatives `r'(variables)`.
+pub trait Constraint: sealed::ConstraintInner {
+    // Note: currently, a constraint's geometric valency is equal to the number of residual
+    // expressions it provides. This might not always be the case, depending on exactly how a
+    // constraint can be specified numerically.
+    /// The valency of this constraint.
     ///
-    /// This tests whether that approximation actually holds.
-    fn test_first_finite_difference<const N: usize>(
-        residual_and_gradient: impl Fn(&[f64; N]) -> (f64, [f64; N]),
-        variable_and_delta_map: impl Fn([f64; N], [f64; N]) -> ([f64; N], [f64; N]),
-    ) {
-        const RELATIVE_EPSILON: f64 = 1e-3;
-        let mut rng = Rng::from_seed(42);
+    /// This is the number of degrees of freedom taken away by the constraint.
+    const VALENCY: u8;
+}
 
-        for _ in 0..5 {
-            let (variables, delta) =
-                variable_and_delta_map(next_f64s(&mut rng), next_f64s(&mut rng));
-            let (r, gradient) = residual_and_gradient(&variables);
-            let (r_plus, _) = residual_and_gradient(&add(variables, delta));
+impl Constraint for PointPointDistance {
+    const VALENCY: u8 = 1;
+}
 
-            let linearized_diff = dot(gradient, delta);
-            let first_finite_diff = r_plus - r;
+impl Constraint for PointPointPointAngle {
+    const VALENCY: u8 = 1;
+}
 
-            assert!(
-                (linearized_diff - first_finite_diff).abs()
-                    / f64::max(linearized_diff.abs(), first_finite_diff.abs())
-                    < RELATIVE_EPSILON,
-                "Difference predicted based on linearized first derivatives and actual first finite difference do not match.\n\
-                Variables: {variables:?}\n\
-                Delta: {delta:?}\n\
-                Predicted: {linearized_diff:?}\n\
-                Actual: {first_finite_diff:?}"
-            );
-        }
-    }
+impl Constraint for PointLineIncidence {
+    const VALENCY: u8 = 1;
+}
 
-    #[test]
-    fn point_point_distance_first_finite_difference() {
-        test_first_finite_difference(
-            |variables| PointPointDistance::compute_residual_and_gradient_(variables, 0.5e0),
-            |variables, delta| {
-                (
-                    variables.map(|d| (d - 0.5) * 1e0),
-                    delta.map(|d| (d - 0.5) * 1e-4),
-                )
-            },
-        );
-        test_first_finite_difference(
-            |variables| PointPointDistance::compute_residual_and_gradient_(variables, 0.5e-9),
-            |variables, delta| {
-                (
-                    variables.map(|d| (d - 0.5) * 1e-10),
-                    delta.map(|d| (d - 0.5) * 1e-14),
-                )
-            },
-        );
-        test_first_finite_difference(
-            |variables| PointPointDistance::compute_residual_and_gradient_(variables, 0.5e10),
-            |variables, delta| {
-                (
-                    variables.map(|d| (d - 0.5) * 1e10),
-                    delta.map(|d| (d - 0.5) * 1e6),
-                )
-            },
-        );
-    }
+impl Constraint for LineLineAngle {
+    const VALENCY: u8 = 1;
+}
 
-    #[test]
-    fn point_point_point_angle_first_finite_difference() {
-        test_first_finite_difference(
-            |variables| {
-                PointPointPointAngle::compute_residual_and_gradient_(variables, 10_f64.to_radians())
-            },
-            |variables, delta| {
-                (
-                    variables.map(|d| (d - 0.5) * 1e0),
-                    delta.map(|d| (d - 0.5) * 1e-5),
-                )
-            },
-        );
-        test_first_finite_difference(
-            |variables| {
-                PointPointPointAngle::compute_residual_and_gradient_(
-                    variables,
-                    -40_f64.to_radians(),
-                )
-            },
-            |variables, delta| {
-                (
-                    variables.map(|d| (d - 0.5) * 1e-10),
-                    delta.map(|d| (d - 0.5) * 1e-15),
-                )
-            },
-        );
-        test_first_finite_difference(
-            |variables| PointPointDistance::compute_residual_and_gradient_(variables, 0.5e10),
-            |variables, delta| {
-                (
-                    variables.map(|d| (d - 0.5) * 1e10),
-                    delta.map(|d| (d - 0.5) * 1e6),
-                )
-            },
-        );
-    }
+impl Constraint for LineLineParallelism {
+    const VALENCY: u8 = 1;
+}
 
-    #[test]
-    fn point_line_incidence_first_finite_difference() {
-        test_first_finite_difference(
-            PointLineIncidence::compute_residual_and_gradient_,
-            |variables, delta| {
-                (
-                    variables.map(|d| (d - 0.5) * 1e0),
-                    delta.map(|d| (d - 0.5) * 1e-4),
-                )
-            },
-        );
-        test_first_finite_difference(
-            PointLineIncidence::compute_residual_and_gradient_,
-            |variables, delta| {
-                (
-                    variables.map(|d| (d - 0.5) * 1e-10),
-                    delta.map(|d| (d - 0.5) * 1e-14),
-                )
-            },
-        );
-    }
-
-    #[test]
-    fn line_line_angle_first_finite_difference() {
-        test_first_finite_difference(
-            |variables| {
-                LineLineAngle::compute_residual_and_gradient_(variables, 10_f64.to_radians())
-            },
-            |variables, delta| {
-                (
-                    variables.map(|d| (d - 0.5) * 1e0),
-                    delta.map(|d| (d - 0.5) * 1e-4),
-                )
-            },
-        );
-        test_first_finite_difference(
-            |variables| {
-                LineLineAngle::compute_residual_and_gradient_(variables, -40_f64.to_radians())
-            },
-            |variables, delta| {
-                (
-                    variables.map(|d| (d - 0.5) * 1e-10),
-                    delta.map(|d| (d - 0.5) * 1e-14),
-                )
-            },
-        );
-    }
-
-    #[test]
-    fn line_line_parallelism_first_finite_difference() {
-        test_first_finite_difference(
-            LineLineParallelism::compute_residual_and_gradient_,
-            |variables, delta| {
-                (
-                    variables.map(|d| (d - 0.5) * 1e0),
-                    delta.map(|d| (d - 0.5) * 1e-4),
-                )
-            },
-        );
-        test_first_finite_difference(
-            LineLineParallelism::compute_residual_and_gradient_,
-            |variables, delta| {
-                (
-                    variables.map(|d| (d - 0.5) * 1e-10),
-                    delta.map(|d| (d - 0.5) * 1e-14),
-                )
-            },
-        );
-    }
-
-    #[test]
-    fn line_circle_tangency_first_finite_difference() {
-        test_first_finite_difference(
-            LineCircleTangency::compute_residual_and_gradient_,
-            |variables, delta| {
-                (
-                    variables.map(|d| (d - 0.5) * 1e0),
-                    delta.map(|d| (d - 0.5) * 1e-4),
-                )
-            },
-        );
-        test_first_finite_difference(
-            LineCircleTangency::compute_residual_and_gradient_,
-            |variables, delta| {
-                (
-                    variables.map(|d| (d - 0.5) * 1e-10),
-                    delta.map(|d| (d - 0.5) * 1e-14),
-                )
-            },
-        );
-    }
+impl Constraint for LineCircleTangency {
+    const VALENCY: u8 = 1;
 }
